separate scanner api module from interfaces
